opengl: core 3.3 fixes

+++ src/modules/graphic/opengl/shaderSources/v130/compute/defren/plan/render_task_subinstgroup_pass2.glsl
@@ -25,7 +25,7 @@ uniform uint pLaneSize;
 layout( local_size_x=64 ) in;
 
 
-const uint cMaxLaneSize = 128;
+const uint cMaxLaneSize = uint( 128 );
 
 shared uint vGroups[ cMaxLaneSize ];
